ref: DetailedView -> slate | proposal
