+++ client/components/stories/Withdraw.stories.tsx
@@ -1,5 +1,12 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import Withdraw from '../../pages/Withdraw';
+import EthereumProvider from '../EthereumProvider';
 
-storiesOf('Withdraw', module).add('Withdraw', () => <Withdraw query="1" />);
+storiesOf('Withdraw', module).add('Withdraw', () => {
+  return (
+    <EthereumProvider>
+      <Withdraw query="1" />;
+    </EthereumProvider>
+  );
+});
