+++ client/components/stories/data.tsx
@@ -7,7 +7,7 @@ import '../../globalStyles.css';
 export const mockedRouter = {
   push: () => {},
   prefetch: () => {},
-  route: '/DetailedView?id=0',
+  route: '/slates/slate?id=0',
   query: '?slateID=0',
   asPath: '/slates/0',
 };
@@ -15,22 +15,6 @@ export const mockedRouter = {
 
 export const epochStartDate = 1549040401;
 export const currentBallot = ballotDates(15499990);
-
-export const unstakedSlate: ISlate = {
-  id: 0,
-  category: 'GRANT',
-  status: statuses.PENDING_TOKENS,
-  deadline: currentBallot.votingOpenDate,
-  title: 'Some slate',
-  owner: 'John Doe',
-  recommenderAddress: '0xd115bffabbdd893a6f7cea402e7338643ced44a6',
-  organization: 'Team Recommender',
-  description:
-    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer eu nibh molestie, auctor ligula a, faucibus ante. Morbi dapibus enim in vulputate congue. Mauris feugiat gravida nibh, sed pellentesque eros pellentesque eu. Sed rutrum vitae magna sed aliquet. Suspendisse facilisis vulputate lobortis. Vestibulum sed dolor eu mi molestie pharetra. Duis ut diam aliquam, molestie erat non, scelerisque ligula. Curabitur accumsan ipsum pellentesque posuere ornare. Sed vulputate cursus accumsan. Morbi efficitur dictum magna, a imperdiet mauris aliquet vitae.',
-  proposals: [],
-  requiredStake: '300000000000000000000',
-  verifiedRecommender: false,
-};
 
 export const proposals: IProposal[] = [
   {
@@ -46,3 +30,19 @@ export const proposals: IProposal[] = [
     teamBackgrounds: 'backgrounds',
   },
 ];
+
+export const unstakedSlate: ISlate = {
+  id: 0,
+  category: 'GRANT',
+  status: 0,
+  deadline: currentBallot.votingOpenDate,
+  title: 'Some slate',
+  owner: 'John Doe',
+  recommenderAddress: '0xd115bffabbdd893a6f7cea402e7338643ced44a6',
+  organization: 'Team Recommender',
+  description:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer eu nibh molestie, auctor ligula a, faucibus ante. Morbi dapibus enim in vulputate congue. Mauris feugiat gravida nibh, sed pellentesque eros pellentesque eu. Sed rutrum vitae magna sed aliquet. Suspendisse facilisis vulputate lobortis. Vestibulum sed dolor eu mi molestie pharetra. Duis ut diam aliquam, molestie erat non, scelerisque ligula. Curabitur accumsan ipsum pellentesque posuere ornare. Sed vulputate cursus accumsan. Morbi efficitur dictum magna, a imperdiet mauris aliquet vitae.',
+  proposals,
+  requiredStake: '300000000000000000000',
+  verifiedRecommender: false,
+};
